Watching changes in read-only transaction
Previously Marmot only relied on simple file watching in order to
capture changes. This was problematic with tools like PocketBase
that would cause no changes to show up and rely on polling to
capture any changes that were not reflected in DB or WAL. By
starting read transaction SQLite will enforce the WAL
changes to be reflected in watch events.